% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance_traveled.R
\name{distance_traveled}
\alias{distance_traveled}
\title{Calculate the `Distance Travelled` by the Multivariable System}
\usage{
distance_traveled(dataIn, group.vars = c("variable"), derivs = T,
  print = T)
}
\arguments{
\item{dataIn}{A data frame containing columns `value`, `time`, and any number of state variables (the order of all columns is irrelevant).}

\item{group.vars}{Character vector of state variable(s) (e.g., species) to be used as grouping variables. Distances will be calculated indiviudally over time, according to these grouping variables. Default = 'variable'}

\item{derivs}{Logical (Default = T), calculate derivatives of distance traveled?}

\item{value}{The values over which distance will be calculated (e.g., species abundances).}
}
\description{
Calculate the `Distance Travelled` by the Multivariable System
}
\examples{
time <- c(1:20)
value <- rnorm(n=20)
variable <- rep(c('a','b'), 20)
dist <- distance_traveled(data.frame(time, variable, value))
plot( dist$time, dist$s) # Cumulative distance traveled
plot( dist$time, dist$dsdt)  # Velocity = First derivative of s (Ds/t or dsdt or s')
plot( dist$time, dist$d2sdt2)  # Acceleration = Second derivative of s (Ds/t or dsdt or s')
}
